# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kudesnik. Windows \u0026 Doors | REST-API scheme",
        "title": "Kudesnik API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список заявок, есть фильтры.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-requests"
                ],
                "summary": "Получить список заявок",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "фильтр по ID админа, обрабатывающего заявку",
                        "name": "admin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "фильтр по статусу заявок",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "фильтр по номеру клиента в формате 79998887766",
                        "name": "customer",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "лимит возвращаемых заявок",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение возвращаемых заявок",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "полученные заявки",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestForList"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать заявку клиента (не требует авторизации)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-requests"
                ],
                "summary": "Создать заявку клиента",
                "parameters": [
                    {
                        "description": "CreateCustomerRequest тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной заявки",
                        "schema": {
                            "$ref": "#/definitions/internal_features_customer_requests_transport_http.CustomerRequestIDResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить заявку по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-requests"
                ],
                "summary": "Получить заявку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID получаемой заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "полученная заявка",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestDetailed"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить заявку по ID. Может только superadmin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-requests"
                ],
                "summary": "Удалить заявку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID удаляемой заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Изменить заявку (желаемые дату и время, доп. комментарий) в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-requests"
                ],
                "summary": "Изменить заявку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID обновляемой заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchCustomerRequest тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestPatch"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_phone_number": {
                    "type": "string"
                },
                "desired_date": {
                    "type": "string"
                },
                "desired_time": {
                    "type": "string"
                },
                "extra_comment": {
                    "type": "string"
                },
                "handled_at": {
                    "type": "string"
                },
                "handler_admin_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestStatus"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestDetailed": {
            "type": "object",
            "properties": {
                "chosen_product": {
                    "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBaseDetailed"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_fullname": {
                    "type": "string"
                },
                "customer_phone_number": {
                    "type": "string"
                },
                "desired_date": {
                    "type": "string"
                },
                "desired_time": {
                    "type": "string"
                },
                "extra_comment": {
                    "type": "string"
                },
                "handled_at": {
                    "type": "string"
                },
                "handler_admin_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestStatus"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestForList": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_fullname": {
                    "type": "string"
                },
                "customer_phone_number": {
                    "type": "string"
                },
                "desired_date": {
                    "type": "string"
                },
                "desired_time": {
                    "type": "string"
                },
                "handled_at": {
                    "type": "string"
                },
                "handler_admin_id": {
                    "type": "string"
                },
                "handler_admin_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestStatus"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestPatch": {
            "type": "object",
            "properties": {
                "desired_date": {
                    "type": "string"
                },
                "desired_time": {
                    "type": "string"
                },
                "extra_comment": {
                    "type": "string"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.CustomerRequestStatus": {
            "type": "string",
            "enum": [
                "new",
                "in_progress",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "NewRequestStatus",
                "InProgressRequestStatus",
                "CompletedRequestStatus",
                "CancelledRequestStatus"
            ]
        },
...
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBaseDetailed": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "installation_price": {
                    "type": "number"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "producer_company_name": {
                    "type": "string"
                },
                "producer_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_customer_requests_transport_http.CustomerRequestIDResponseDTO": {
            "type": "object",
            "properties": {
                "customer_request_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Some queries require authorization: \"Bearer \u003cjwt_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}